Explicitly include sipbuild/module/source in the built tree
When building from a git repository checkout, setuptools_scm takes
care of including all files that are present in the git repository.

However, when building from an unpacked archive, like we do in Debian,
that is not the case, and without this patch sipbuild/module/source
was not included in the resulting .deb package.