Added the `%MinimumABIVersion` directive
The `%MinimumABIVersion` directive is used to specify the minimum `sip`
module ABI to target in order to ensure that all calls to the public API
from handwritten code are available.

A consequence of this change is that passing string values to the
`/Deprecated/` annotation no longer requires an ABI version that supports
them to be targeted.  If the targeted ABI version does not support them
they are simply ignored.

Resolves #65