`sip` Module Configuration
The `%SipModuleConfiguration` directive is used by ABI v14 to specify how
the `sip` module is expected to behave.  For example it can be configured
to behave like v12 or v13.

The `--option` option has been added to the `sip-module` command line tool
in order to configure a generated sdist to behave accordingly.